refactor: groups로 모듈 그룹핑
