+++ core/navigation/build.gradle.kts
@@ -12,10 +12,12 @@ dependencies {
     implementation(projects.feature.canvas.api)
     implementation(projects.feature.gallery.api)
     implementation(projects.feature.login.api)
-    implementation(projects.feature.grouphome.api)
     implementation(projects.feature.segmentation.api)
     implementation(projects.feature.splash.api)
-    implementation(projects.feature.groupenter.api)
+
+    implementation(projects.feature.groups.enter.api)
+    implementation(projects.feature.groups.home.api)
+    implementation(projects.feature.groups.list.api)
 
     implementation(libs.bundles.navigation)
 }
