--- feature/groupenter/api/src/main/java/com/teamyg/parfait/feature/groupenter/api/NavKeyGroupInviteCode.kt
@@ -1,7 +1,0 @@
-package com.teamyg.parfait.feature.groupenter.api
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object NavKeyGroupInviteCode : NavKey
