--- feature/groupenter/impl/src/main/java/com/teamyg/parfait/feature/groupenter/impl/NavigationModule.kt
@@ -1,20 +1,0 @@
-package com.teamyg.parfait.feature.groupenter.impl
-
-import androidx.navigation3.runtime.EntryProviderScope
-import androidx.navigation3.runtime.NavKey
-import com.teamyg.parfait.core.navigation.Navigator
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityRetainedComponent
-import dagger.multibindings.IntoSet
-
-@Module
-@InstallIn(ActivityRetainedComponent::class)
-object NavigationModule {
-    @IntoSet
-    @Provides
-    fun provideFeatureGroupInviteCodeEntryBuilder(): EntryProviderScope<NavKey>.(Navigator) -> Unit = {
-        featureGroupInviteCodeEntryBuilder(navigator = it)
-    }
-}
