--- feature/grouphome/api/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    alias(libs.plugins.parfait.module.feature.api)
-}
-
-android {
-    namespace = "com.teamyg.parfait.feature.grouphome.api"
-}
