--- feature/grouphome/api/src/main/java/com/teamyg/parfait/feature/grouphome/api/NavKeyGroupHome.kt
@@ -1,7 +1,0 @@
-package com.teamyg.parfait.feature.grouphome.api
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class NavKeyGroupHome(val groupId: Int) : NavKey
