--- feature/grouphome/impl/src/main/java/com/teamyg/parfait/feature/grouphome/impl/EntryBuilder.kt
@@ -1,24 +1,0 @@
-package com.teamyg.parfait.feature.grouphome.impl
-
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.ui.Modifier
-import androidx.navigation3.runtime.EntryProviderScope
-import androidx.navigation3.runtime.NavKey
-import com.teamyg.parfait.feature.grouphome.api.NavKeyGroupHome
-import com.teamyg.parfait.core.navigation.Navigator
-
-fun EntryProviderScope<NavKey>.featureGroupHomeEntryBuilder(navigator: Navigator) {
-    entry<NavKeyGroupHome> { key ->
-        Scaffold { innerPadding ->
-            GroupHomeRoute(
-                navigator = navigator,
-                key = key,
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(innerPadding),
-            )
-        }
-    }
-}
