+++ feature/groups/enter/api/src/main/AndroidManifest.xml
