+++ feature/groups/enter/api/src/main/kotlin/com/teamyg/parfait/feature/groups/enter/api/NavKeyGroupInviteCode.kt
@@ -1,0 +1,7 @@
+package com.teamyg.parfait.feature.groups.enter.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object NavKeyGroupInviteCode : NavKey
