+++ feature/groups/enter/impl/src/main/AndroidManifest.xml
