+++ feature/groups/enter/impl/src/main/res/drawable/ic_chevron_left.xml
