+++ feature/groups/home/api/src/main/AndroidManifest.xml
