+++ feature/groups/home/api/src/main/kotlin/com/teamyg/parfait/feature/groups/home/api/NavKeyGroupHome.kt
@@ -1,0 +1,7 @@
+package com.teamyg.parfait.feature.groups.home.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class NavKeyGroupHome(val groupId: Int) : NavKey
