+++ feature/groups/home/impl/src/main/AndroidManifest.xml
