+++ settings.gradle.kts
@@ -43,8 +43,6 @@ include(":domain")
 include(
     ":feature:login:api",
     ":feature:login:impl",
-    ":feature:grouphome:api",
-    ":feature:grouphome:impl",
     ":feature:segmentation:api",
     ":feature:segmentation:impl",
     ":feature:camera:api",
@@ -55,6 +53,13 @@ include(
     ":feature:canvas:impl",
     ":feature:splash:api",
     ":feature:splash:impl",
-    ":feature:groupenter:api",
-    ":feature:groupenter:impl",
+)
+
+include(
+    ":feature:groups:enter:api",
+    ":feature:groups:enter:impl",
+    ":feature:groups:home:api",
+    ":feature:groups:home:impl",
+    ":feature:groups:list:api",
+    ":feature:groups:list:impl",
 )
